renaming gui signals file
